Add icon utility, modify color utility, and add typography variables
This commit introduces several changes in the style utilities and variables. A new utility for icons was added with default display set to block. The color utility was renamed to properties and extended with font-size properties and animation duration. The file `src/styles/abstract/variables/_index.scss` was updated to include new typography variables. In the file `src/styles/abstract/variables/_colors.scss`, a new color was defined in the color palette. Lastly, the global font color was modified in `src/styles.scss`. These changes aim to improve component styling flexibility and maintainability.